New Scene and Enemy Prototype
Updated to include the changes that Leo made. DON'T TOUCH THE CAMERA --
unless you know how to fix it. I still have to figure out a solution for
the resolution problem. If you want to make a new scene I suggest
duplicating Lvl3 so that all of the settings will already be in place.