Added GameOver
I have added a new scene to the end of the game.  GameOver.  I also
added the functionality to the playercontroller script to where if you
die it goes to GameOver instead of reloading the game.  Also, I modified
a few other minor things to make this fit.  I do have the Splash screen
for "Team Dungeon Masters" and "Made With Unity" if anyone wants to
check them out and comment on them.