Removed/added scripts
I have added a new script that is attached to the intro splash screens
to automatically advance after 10 seconds or on an escape button press.
Also, I have added the splash screens to the list of scenes and modified
a few lines of code in playercontroller and score to allow for them.
Also I removed the script to auto re-size the planes in the splash pages
and game over scenes.